+++ src/main/java/roomescape/reservation/Reservation.java
@@ -1,7 +1,6 @@
 package roomescape.reservation;
 
 import java.time.LocalDate;
-import roomescape.time.ReservationTime;
 
 public class Reservation {
     private final Long id;
