+++ src/main/java/roomescape/reservation/web/ReservationController.java
@@ -13,8 +13,10 @@ import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import roomescape.reservation.Reservation;
 import roomescape.reservation.dao.ReservationDao;
-import roomescape.time.ReservationTime;
-import roomescape.time.dao.ReservationTimeDao;
+import roomescape.reservation.ReservationTime;
+import roomescape.reservation.dao.ReservationTimeDao;
+import roomescape.reservation.web.dto.ReservationRequest;
+import roomescape.reservation.web.dto.ReservationResponse;
 
 @Controller
 @RequestMapping("/reservations")
