+++ src/test/java/roomescape/reservation/ReservationTest.java
@@ -10,7 +10,6 @@ import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.NullAndEmptySource;
 import org.junit.jupiter.params.provider.ValueSource;
-import roomescape.time.ReservationTime;
 
 class ReservationTest {
     @DisplayName("id가_같은지_여부를_반환한다")
